Move elasticsearch highlighting code out of results type
